+++ subscriber/podaac_data_downloader.py
@@ -203,6 +203,14 @@ def cmr_downloader(args, token, data_path):
             'from -dc, -dy, -dydoy, or -dymd'
         )
 
+    if args.subset and args.search_cycles:
+        # Cycle+Subset are not supported, because Harmony does not
+        # currently accept Cycle.
+        raise ValueError(
+            'Error: Incompatible Parameters. You\'ve provided both cycles and subset, which is '
+            'not allowed. Please provide either cycles or subset separately, but not both.'
+        )
+
     if args.offset:
         ts_shift = timedelta(hours=int(args.offset))